+++ src/components/settings/MiningFeesFormSchema.ts
